+++ src/Framework/src/Routing/Commands/RouteListCommand.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 namespace Aphiria\Framework\Routing\Commands;
 
 use Aphiria\Console\Commands\Command;
+use Aphiria\Console\Input\Option;
+use Aphiria\Console\Input\OptionType;
 
 /**
  * Defines the command to list your application's routes
@@ -21,6 +23,10 @@ final class RouteListCommand extends Command
 {
     public function __construct()
     {
-        parent::__construct('route:list', [], [], 'Lists the routes in your app');
+        $options = [
+            new Option('fqn', OptionType::NoValue, description: 'Shows the fully-qualified class names of controllers and middleware'),
+            new Option('middleware', OptionType::NoValue, description: 'Shows the middleware for each route')
+        ];
+        parent::__construct('route:list', [], $options, 'Lists the routes in your app');
     }
 }