+++ src/Framework/tests/Routing/Commands/Mocks/MiddlewareA.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * Aphiria
+ *
+ * @link      https://www.aphiria.com
+ * @copyright Copyright (C) 2022 David Young
+ * @license   https://github.com/aphiria/aphiria/blob/1.x/LICENSE.md
+ */
+
+declare(strict_types=1);
+
+namespace Aphiria\Framework\Tests\Routing\Commands\Mocks;
+
+/**
+ * Mocks a middleware class
+ */
+class MiddlewareA
+{
+}