+++ src/Framework/tests/Routing/Commands/RouteListCommandHandlerTest.php
@@ -17,7 +17,10 @@ use Aphiria\Console\Output\Formatters\PaddingFormatter;
 use Aphiria\Console\Output\IOutput;
 use Aphiria\Console\StatusCode;
 use Aphiria\Framework\Routing\Commands\RouteListCommandHandler;
+use Aphiria\Framework\Tests\Routing\Commands\Mocks\MiddlewareA;
+use Aphiria\Framework\Tests\Routing\Commands\Mocks\MiddlewareB;
 use Aphiria\Routing\Matchers\Constraints\HttpMethodRouteConstraint;
+use Aphiria\Routing\Middleware\MiddlewareBinding;
 use Aphiria\Routing\Route;
 use Aphiria\Routing\RouteAction;
 use Aphiria\Routing\RouteCollection;
@@ -43,6 +46,36 @@ class RouteListCommandHandlerTest extends TestCase
         $this->output = $this->createMock(IOutput::class);
     }
 
+    /**
+     * Gets URI templates for testing
+     *
+     * @return list<array{0: string, 1: string}> The list of URI templates and expected outputs
+     */
+    public function getUriTemplates(): array
+    {
+        return [
+            ['/foo', '/foo'],
+            ['/foo/:bar', '/foo/<info>:bar</info>'],
+            ['/foo/:bar/baz', '/foo/<info>:bar</info>/baz'],
+            ['/foo/:bar(int)', '/foo/<info>:bar(int)</info>'],
+            ['/foo[/:bar[/:baz]]', '/foo[/<info>:bar</info>[/<info>:baz</info>]]']
+        ];
+    }
+
+    public function testFullyQualifiedClassNamesAreUsedForControllersAndMiddlewareWhenOptionIsSpecified(): void
+    {
+        $route = new Route(
+            new UriTemplate(''),
+            new RouteAction(self::class, 'foo'),
+            [new HttpMethodRouteConstraint(['POST'])],
+            [new MiddlewareBinding(MiddlewareA::class)]
+        );
+        $this->routes->add($route);
+        $this->setUpOutputExpectations([['POST', '/', '<comment>' . self::class . '::foo</comment>'], ['', '', '↳ ' . MiddlewareA::class]]);
+        $input = new Input('route:list', [], ['fqn' => null, 'middleware' => null]);
+        $this->assertSame(StatusCode::Ok, $this->commandHandler->handle($input, $this->output));
+    }
+
     public function testHandlingRouteWithoutHttpMethodConstraintThrowsException(): void
     {
         $this->expectException(RuntimeException::class);
@@ -56,12 +89,49 @@ class RouteListCommandHandlerTest extends TestCase
         $route = new Route(
             new UriTemplate(''),
             new RouteAction(self::class, 'foo'),
-            // Purposely registering out of order
+            // Purposely registering out of alphabetic order
             [new HttpMethodRouteConstraint(['POST', 'DELETE'])]
         );
         $this->routes->add($route);
-        $this->setUpOutputExpectations([['DELETE|POST', '/', self::class . '::foo']]);
+        $this->setUpOutputExpectations([['DELETE|POST', '/', '<comment>RouteListCommandHandlerTest::foo</comment>']]);
         $this->assertSame(StatusCode::Ok, $this->commandHandler->handle($this->input, $this->output));
+    }
+
+    public function testMiddlewareOptionConcatenatesMiddlewareClassNamesBelowRouteData(): void
+    {
+        $route = new Route(
+            new UriTemplate(''),
+            new RouteAction(self::class, 'foo'),
+            [new HttpMethodRouteConstraint(['POST'])],
+            [new MiddlewareBinding(MiddlewareA::class), new MiddlewareBinding(MiddlewareB::class)]
+        );
+        $this->routes->add($route);
+        $this->setUpOutputExpectations([['POST', '/', '<comment>RouteListCommandHandlerTest::foo</comment>'], ['', '', '↳ MiddlewareA → MiddlewareB']]);
+        $input = new Input('route:list', [], ['middleware' => null]);
+        $this->assertSame(StatusCode::Ok, $this->commandHandler->handle($input, $this->output));
+    }
+
+    public function testMiddlewareOptionDoesNotAddRowForRoutesWithNoMiddleware(): void
+    {
+        $route1 = new Route(
+            new UriTemplate('/bar'),
+            new RouteAction(self::class, 'bar'),
+            [new HttpMethodRouteConstraint(['POST'])]
+        );
+        $route2 = new Route(
+            new UriTemplate('/foo'),
+            new RouteAction(self::class, 'foo'),
+            [new HttpMethodRouteConstraint(['POST'])],
+            [new MiddlewareBinding(MiddlewareA::class), new MiddlewareBinding(MiddlewareB::class)]
+        );
+        $this->routes->addMany([$route1, $route2]);
+        $this->setUpOutputExpectations([
+            ['POST', '/bar', '<comment>RouteListCommandHandlerTest::bar</comment>'],
+            ['POST', '/foo', '<comment>RouteListCommandHandlerTest::foo</comment>'],
+            ['', '', '↳ MiddlewareA → MiddlewareB']
+        ]);
+        $input = new Input('route:list', [], ['middleware' => null]);
+        $this->assertSame(StatusCode::Ok, $this->commandHandler->handle($input, $this->output));
     }
 
     public function testRoutesAreAlphabetizedByPath(): void
@@ -79,8 +149,8 @@ class RouteListCommandHandlerTest extends TestCase
         );
         $this->routes->addMany([$route1, $route2]);
         $this->setUpOutputExpectations([
-            ['POST', '/bar', self::class . '::bar'],
-            ['POST', '/foo', self::class . '::foo']
+            ['POST', '/bar', '<comment>RouteListCommandHandlerTest::bar</comment>'],
+            ['POST', '/foo', '<comment>RouteListCommandHandlerTest::foo</comment>']
         ]);
         $this->assertSame(StatusCode::Ok, $this->commandHandler->handle($this->input, $this->output));
     }
@@ -100,10 +170,42 @@ class RouteListCommandHandlerTest extends TestCase
         );
         $this->routes->addMany([$route1, $route2]);
         $this->setUpOutputExpectations([
-            ['DELETE', '/foo', self::class . '::bar'],
-            ['POST', '/foo', self::class . '::foo']
+            ['DELETE', '/foo', '<comment>RouteListCommandHandlerTest::bar</comment>'],
+            ['POST', '/foo', '<comment>RouteListCommandHandlerTest::foo</comment>']
         ]);
         $this->assertSame(StatusCode::Ok, $this->commandHandler->handle($this->input, $this->output));
+    }
+
+    /**
+     * @dataProvider getUriTemplates
+     *
+     * @param string $uriTemplate The raw URI template
+     * @param string $expectedFormattedUriTemplate The expected formatted URI template
+     */
+    public function testRouteVariablesInUriTemplatesAreHighlighted(string $uriTemplate, string $expectedFormattedUriTemplate): void
+    {
+        $route = new Route(
+            new UriTemplate($uriTemplate),
+            new RouteAction(self::class, 'foo'),
+            [new HttpMethodRouteConstraint(['POST'])]
+        );
+        $this->routes->add($route);
+        $this->setUpOutputExpectations([['POST', $expectedFormattedUriTemplate, '<comment>RouteListCommandHandlerTest::foo</comment>']]);
+        $this->assertSame(StatusCode::Ok, $this->commandHandler->handle($this->input, $this->output));
+    }
+
+    public function testShortQualifiedClassNamesAreUsedForControllersAndMiddlewareWhenOptionIsSpecified(): void
+    {
+        $route = new Route(
+            new UriTemplate(''),
+            new RouteAction(self::class, 'foo'),
+            [new HttpMethodRouteConstraint(['POST'])],
+            [new MiddlewareBinding(MiddlewareA::class)]
+        );
+        $this->routes->add($route);
+        $input = new Input('route:list', [], ['middleware' => null]);
+        $this->setUpOutputExpectations([['POST', '/', '<comment>RouteListCommandHandlerTest::foo</comment>'], ['', '', '↳ MiddlewareA']]);
+        $this->assertSame(StatusCode::Ok, $this->commandHandler->handle($input, $this->output));
     }
 
     /**
@@ -114,7 +216,7 @@ class RouteListCommandHandlerTest extends TestCase
     private function setUpOutputExpectations(array $expectedRows): void
     {
         // Prepend the header row
-        $expectedRows = [['Method', 'Path', 'Action'], ...$expectedRows];
+        $expectedRows = [['<b>Method</b>', '<b>Path</b>', '<b>Action</b>'], ...$expectedRows];
 
         $this->output->expects($this->once())
             ->method('writeln')